TCPCL: Refactored Client state into own file
